#fix #jsx Updated components to use TableComponent instead of Table from antd and made necessary changes to display data correctly and improve styling. Refactored file display. Refactored the alignment and style in the rendering methods of the HighSchoolData component.
