+++ frontend/src/modules/management.jsx
@@ -1,0 +1,503 @@
+import ButtonComponent from "@/components/button/index.jsx";
+import IconComponent from "@/components/icons/index.jsx";
+import { Paragraph } from "@/components/title/index.jsx";
+import ColorsContext from "@/context/colors.jsx";
+import { CheckProgress } from "@/modules/progress.jsx";
+import {
+  DeleteOutlined,
+  ExportOutlined,
+  FormOutlined,
+} from "@ant-design/icons";
+import { useContext } from "react";
+import { GoClock, GoEye } from "react-icons/go";
+import { TbActivityHeartbeat } from "react-icons/tb";
+
+export const LocationModule = () => {
+  const columns = [
+    {
+      title: "Location name",
+      key: "name",
+      dataIndex: "name",
+      render: (text) => {
+        return (
+          <Paragraph className={"text-start"} fontSize={"text-lg"}>
+            {text}
+          </Paragraph>
+        );
+      },
+    },
+    {
+      title: "Location code",
+      key: "code",
+      dataIndex: "code",
+      align: "center",
+      render: (text) => {
+        return (
+          <Paragraph className={"text-center"} fontSize={"text-lg"}>
+            {text}
+          </Paragraph>
+        );
+      },
+    },
+    {
+      title: "Type",
+      key: "type",
+      dataIndex: "type",
+      align: "center",
+      render: (text) => {
+        return (
+          <Paragraph className={"text-center"} fontSize={"text-lg"}>
+            {text}
+          </Paragraph>
+        );
+      },
+    },
+    {
+      title: "Status",
+      key: "status",
+      dataIndex: "status",
+      align: "center",
+      render: (text) => {
+        const { bg, hover } = CheckProgress(text);
+        return (
+          <ButtonComponent
+            defaultBg={bg}
+            defaultHoverBg={hover}
+            controlHeight={50}
+            borderRadius={5}
+            style={{ width: 128 }}
+          >
+            {text?.toUpperCase()}
+          </ButtonComponent>
+        );
+      },
+    },
+    {
+      title: "Action",
+      key: "action",
+      align: "center",
+      render: () => (
+        <div className={" space-x-2.5 "}>
+          <IconComponent
+            className={"text-xl text-indigo-500 border border-indigo-600"}
+            style={{
+              borderRadius: 5,
+              paddingLeft: 4,
+              paddingRight: 4,
+            }}
+            icon={<FormOutlined />}
+          />
+
+          <IconComponent
+            className={"text-xl text-red-600 border border-indigo-600"}
+            style={{
+              borderRadius: 5,
+              paddingLeft: 4,
+              paddingRight: 4,
+            }}
+            icon={<DeleteOutlined />}
+          />
+
+          <IconComponent
+            className={"text-xl text-indigo-500 border border-indigo-600"}
+            style={{
+              borderRadius: 5,
+              paddingLeft: 4,
+              paddingRight: 4,
+            }}
+            icon={<ExportOutlined />}
+          />
+        </div>
+      ),
+    },
+  ];
+
+  const data = [
+    {
+      name: "Mason Location",
+      code: 8581,
+      status: "Active",
+      type: "Main office only",
+      render: (text) => (
+        <Paragraph fontSize={"text-lg"} fontWeightStrong={400}>
+          {text}
+        </Paragraph>
+      ),
+    },
+    {
+      name: "Mason Location",
+      code: 8581,
+      status: "Process",
+      type: "Main office only",
+      render: (text) => (
+        <Paragraph fontSize={"text-lg"} fontWeightStrong={400}>
+          {text}
+        </Paragraph>
+      ),
+    },
+    {
+      name: "Mason Location",
+      code: 8581,
+      status: "Close",
+      type: "Main office only",
+      render: (text) => (
+        <Paragraph fontSize={"text-lg"} fontWeightStrong={400}>
+          {text}
+        </Paragraph>
+      ),
+    },
+  ];
+
+  return { columns, data };
+};
+
+export const HighSchoolModule = () => {
+  const columns = [
+    {
+      title: "School name",
+      key: "name",
+      dataIndex: "name",
+      render: (text) => (
+        <Paragraph fontSize={"text-lg"} fontWeightStrong={400}>
+          {text}
+        </Paragraph>
+      ),
+    },
+    {
+      title: "School code",
+      key: "code",
+      dataIndex: "code",
+      align: "center",
+      render: (text) => {
+        return (
+          <Paragraph className={"text-center"} fontSize={"text-lg"}>
+            {text}
+          </Paragraph>
+        );
+      },
+    },
+    {
+      title: "High school map",
+      key: "map",
+      dataIndex: "map",
+      align: "center",
+      render: () => (
+        <div className={"text-center"}>
+          <IconComponent
+            className={"text-2xl cursor-default"}
+            icon={<GoEye />}
+          />
+        </div>
+      ),
+    },
+    {
+      title: "Status",
+      key: "status",
+      dataIndex: "status",
+      align: "center",
+      render: (text) => {
+        const { bg, hover } = CheckProgress(text);
+        return (
+          <ButtonComponent
+            defaultBg={bg}
+            defaultHoverBg={hover}
+            controlHeight={50}
+            borderRadius={5}
+            style={{ width: 128 }}
+          >
+            {text?.toUpperCase()}
+          </ButtonComponent>
+        );
+      },
+    },
+    {
+      title: "Action",
+      key: "action",
+      align: "center",
+      render: () => (
+        <div className={" space-x-2.5 "}>
+          <IconComponent
+            className={"text-xl text-indigo-500 border border-indigo-600"}
+            style={{
+              borderRadius: 5,
+              paddingLeft: 4,
+              paddingRight: 4,
+            }}
+            icon={<FormOutlined />}
+          />
+
+          <IconComponent
+            className={
+              "text-xl text-indigo-600 hover:text-indigo-600 border border-indigo-600"
+            }
+            icon={<GoClock />}
+            style={{
+              borderRadius: 5,
+              paddingLeft: 4,
+              paddingRight: 4,
+            }}
+          />
+        </div>
+      ),
+    },
+  ];
+
+  const data = [
+    {
+      name: "Mason Location",
+      code: 8581,
+      status: "Active",
+    },
+    {
+      name: "Mason Location",
+      code: 8581,
+      status: "Process",
+    },
+    {
+      name: "Mason Location",
+      code: 8581,
+      status: "Close",
+    },
+  ];
+
+  return { columns, data };
+};
+
+export const HearModule = () => {
+  const columns = [
+    {
+      title: "Name",
+      key: "name",
+      dataIndex: "name",
+      align: "center",
+      render: (text) => {
+        return (
+          <Paragraph className={"text-start"} fontSize={"text-lg"}>
+            {text}
+          </Paragraph>
+        );
+      },
+    },
+    {
+      title: "School code",
+      key: "code",
+      dataIndex: "code",
+      align: "center",
+      render: (text) => {
+        return (
+          <Paragraph className={"text-center"} fontSize={"text-lg"}>
+            {text}
+          </Paragraph>
+        );
+      },
+    },
+    {
+      title: "Status",
+      key: "status",
+      dataIndex: "status",
+      align: "center",
+      render: (text) => {
+        const { bg, hover } = CheckProgress(text);
+        return (
+          <ButtonComponent
+            defaultBg={bg}
+            defaultHoverBg={hover}
+            controlHeight={50}
+            borderRadius={5}
+            style={{ width: 128 }}
+          >
+            {text?.toUpperCase()}
+          </ButtonComponent>
+        );
+      },
+    },
+    {
+      title: "Action",
+      key: "action",
+      align: "center",
+      render: () => (
+        <div className={" space-x-2.5 "}>
+          <IconComponent
+            className={"text-xl text-indigo-500 border border-indigo-600"}
+            style={{
+              borderRadius: 5,
+              paddingLeft: 4,
+              paddingRight: 4,
+            }}
+            icon={<FormOutlined />}
+          />
+
+          <IconComponent
+            className={
+              "text-xl text-indigo-600 hover:text-indigo-600 border border-indigo-600"
+            }
+            icon={<GoClock />}
+            style={{
+              borderRadius: 5,
+              paddingLeft: 4,
+              paddingRight: 4,
+            }}
+          />
+        </div>
+      ),
+    },
+  ];
+
+  const data = [
+    {
+      name: "Facebook",
+      code: 8581,
+      status: "Active",
+    },
+    {
+      name: "Google",
+      code: 8581,
+      status: "Process",
+    },
+    {
+      name: "Other",
+      code: 8581,
+      status: "Close",
+    },
+    {
+      name: "Recommendation",
+      code: 8581,
+      status: "Close",
+    },
+    {
+      name: "Yelp",
+      code: 8581,
+      status: "Active",
+    },
+  ];
+
+  return { columns, data };
+};
+
+export const VehiclesModule = () => {
+  const { colorsObject } = useContext(ColorsContext);
+  const columns = [
+    {
+      title: "Location name",
+      key: "name",
+      dataIndex: "name",
+      render: (text) => {
+        return <Paragraph fontSize={"text-lg"}>{text}</Paragraph>;
+      },
+    },
+    {
+      title: "Location code",
+      key: "code",
+      dataIndex: "code",
+      align: "center",
+      render: (text) => {
+        return (
+          <Paragraph className={"text-center"} fontSize={"text-lg"}>
+            {text}
+          </Paragraph>
+        );
+      },
+    },
+    {
+      title: "Type",
+      key: "type",
+      dataIndex: "type",
+      align: "center",
+      render: (text) => {
+        return (
+          <Paragraph
+            className={"text-center"}
+            fontSize={"text-lg"}
+            fontWeightStrong={400}
+          >
+            {text}
+          </Paragraph>
+        );
+      },
+    },
+    {
+      title: "Appointment color",
+      key: "appointment",
+      dataIndex: "appointment",
+      align: "center",
+      render: (color) => {
+        return (
+          <div
+            className="w-32 m-auto"
+            style={{
+              height: 30,
+              borderRadius: 5,
+              background: color,
+            }}
+          ></div>
+        );
+      },
+    },
+    {
+      title: "Action",
+      key: "action",
+      align: "center",
+      render: () => (
+        <div className={" space-x-2.5 "}>
+          <IconComponent
+            className={"text-xl text-indigo-500 border border-indigo-600"}
+            style={{
+              borderRadius: 5,
+              paddingLeft: 4,
+              paddingRight: 4,
+            }}
+            icon={<FormOutlined />}
+          />
+
+          <IconComponent
+            className={
+              "text-xl text-indigo-600 hover:text-indigo-600 border border-indigo-600"
+            }
+            icon={<TbActivityHeartbeat />}
+            classNames={"items-center"}
+            style={{
+              borderRadius: 5,
+              paddingLeft: 4,
+              paddingRight: 4,
+            }}
+          />
+
+          <IconComponent
+            className={
+              "text-xl text-indigo-600 hover:text-indigo-600 border border-indigo-600"
+            }
+            icon={<GoClock />}
+            style={{
+              borderRadius: 5,
+              paddingLeft: 4,
+              paddingRight: 4,
+            }}
+          />
+        </div>
+      ),
+    },
+  ];
+
+  const data = [
+    {
+      name: "Mason Location",
+      code: 8581,
+      appointment: colorsObject.info,
+      type: "Car",
+    },
+    {
+      name: "Mason Location",
+      code: 8581,
+      appointment: colorsObject.secondary,
+      type: "Car",
+    },
+    {
+      name: "Mason Location",
+      code: 8581,
+      appointment: colorsObject.danger,
+      type: "Car",
+    },
+  ];
+
+  return { columns, data };
+};
