+++ frontend/src/modules/product.jsx
@@ -1,0 +1,178 @@
+import ButtonComponent from "@/components/button/index.jsx";
+import IconComponent from "@/components/icons/index.jsx";
+import { Paragraph } from "@/components/title/index.jsx";
+import { CheckProgress } from "@/modules/progress.jsx";
+import {
+  DeleteOutlined,
+  ExportOutlined,
+  FormOutlined,
+} from "@ant-design/icons";
+import { Space } from "antd";
+
+export const ProductModule = () => {
+  const columns = [
+    {
+      title: "Item Name",
+      dataIndex: "name",
+      key: "name",
+      render: (text) => (
+        <Paragraph fontSize={"text-lg"} fontWeightStrong={400}>
+          {text}
+        </Paragraph>
+      ),
+    },
+    {
+      title: "Code number",
+      dataIndex: "codeNumber",
+      key: "codeNumber",
+      render: (text) => (
+        <div className={"text-center"}>
+          <Paragraph fontSize={"text-lg"} fontWeightStrong={400}>
+            {text}
+          </Paragraph>
+        </div>
+      ),
+    },
+    {
+      title: "Type",
+      dataIndex: "type",
+      key: "type",
+      render: (text) => (
+        <div className={"text-center"}>
+          <Paragraph fontSize={"text-lg"} fontWeightStrong={400}>
+            {text}
+          </Paragraph>
+        </div>
+      ),
+    },
+    {
+      title: "Sub type",
+      dataIndex: "subtype",
+      key: "subtype",
+      render: (text) => (
+        <div className={"text-center"}>
+          <Paragraph fontSize={"text-lg"} fontWeightStrong={400}>
+            {text}
+          </Paragraph>
+        </div>
+      ),
+    },
+    {
+      title: "BTW HOURS",
+      dataIndex: "hours",
+      key: "hours",
+      render: (text) => (
+        <div className={"text-center"}>
+          <Paragraph fontSize={"text-lg"} fontWeightStrong={400}>
+            {text}
+          </Paragraph>
+        </div>
+      ),
+    },
+    {
+      title: "Observation Hours",
+      dataIndex: "observation",
+      key: "observation",
+      render: (text) => (
+        <div className={"text-center"}>
+          <Paragraph fontSize={"text-lg"} fontWeightStrong={400}>
+            {text}
+          </Paragraph>
+        </div>
+      ),
+    },
+    {
+      title: "Status",
+      dataIndex: "status",
+      key: "status",
+      render: (text) => {
+        const { bg, hover } = CheckProgress(text);
+        return (
+          <Space size={"middle"}>
+            <ButtonComponent
+              defaultBg={bg}
+              defaultHoverBg={hover}
+              controlHeight={50}
+              borderRadius={5}
+              style={{ width: 128 }}
+            >
+              {text.toUpperCase()}
+            </ButtonComponent>
+          </Space>
+        );
+      },
+    },
+    {
+      title: "Action",
+      key: "action",
+      render: () => (
+        <div className={"space-x-2.5"}>
+          <IconComponent
+            className={"text-xl text-indigo-500 border border-indigo-600"}
+            style={{
+              borderRadius: 5,
+              paddingLeft: 4,
+              paddingRight: 4,
+            }}
+            icon={<FormOutlined />}
+          />
+
+          <IconComponent
+            className={"text-xl text-red-600 border border-indigo-600"}
+            style={{
+              borderRadius: 5,
+              paddingLeft: 4,
+              paddingRight: 4,
+            }}
+            icon={<DeleteOutlined />}
+          />
+
+          <IconComponent
+            className={"text-xl text-indigo-500 border border-indigo-600"}
+            style={{
+              borderRadius: 5,
+              paddingLeft: 4,
+              paddingRight: 4,
+            }}
+            icon={<ExportOutlined />}
+          />
+        </div>
+      ),
+    },
+  ];
+
+  const data = [
+    {
+      key: "1",
+      name: "Advanced Parking",
+      codeNumber: "001",
+      type: "Owner",
+      subtype: "Teen BTW",
+      hours: 1,
+      observation: 1,
+      status: "Active",
+    },
+    {
+      key: "2",
+      name: "Advanced Parking",
+      codeNumber: "001",
+      type: "Owner",
+      subtype: "Teen BTW",
+      hours: 1,
+      observation: 1,
+      status: "Close",
+    },
+    {
+      key: "2",
+      name: "Advanced Parking",
+      codeNumber: "001",
+      type: "Owner",
+      subtype: "Teen BTW",
+      hours: 1,
+      observation: 1,
+      status: "process",
+    },
+  ];
+
+  return { data, columns };
+};
