+++ frontend/src/modules/quiz-exam.jsx
@@ -1,0 +1,31 @@
+export const ExamModule = () => {
+  const columns = [
+    {
+      title: "Quiz name",
+      dataIndex: "name",
+      key: "name",
+    },
+    {
+      title: "Questions Per Quiz",
+      dataIndex: "questionsPer",
+      key: "questionsPer",
+    },
+    {
+      title: "Question Pool",
+      dataIndex: "questionPool",
+      key: "questionPool",
+    },
+    {
+      title: "Passing grade",
+      dataIndex: "grade",
+      key: "grade",
+    },
+    {
+      title: "Action",
+      dataIndex: "action",
+      key: "action",
+    },
+  ];
+
+  return { columns };
+};
