+++ frontend/src/pages/managment/file/index.jsx
@@ -6,9 +6,12 @@ import {
 } from "@/components/form/index.jsx";
 import IconComponent, { Icons } from "@/components/icons/index.jsx";
 import Modal from "@/components/modal/index.jsx";
+import TableComponent from "@/components/table/index.jsx";
 import Title, { Paragraph } from "@/components/title/index.jsx";
 import ColorsContext from "@/context/colors.jsx";
 import { useDate } from "@/hooks/useDate.jsx";
+import { FileCategoryModule } from "@/modules/file-category.jsx";
+import { FileChart } from "@/pages/managment/file/items/chart.jsx";
 import { Space, Switch, Table } from "antd";
 import classNames from "classnames";
 import { Fragment, useContext, useState } from "react";
@@ -51,85 +54,9 @@ const mockData = [
   { key: "11", title: "Title 5", description: "Sample Description 5" },
 ];
 
-const columns = [
-  {
-    title: "Category",
-    key: "category",
-    dataIndex: "category",
-  },
-  {
-    title: "File status",
-    key: "status",
-    dataIndex: "status",
-    render: (text) => {
-      return (
-        <Space size={"middle"}>
-          <ButtonComponent
-            defaultBg={CheckProgress(text)}
-            defaultHoverBg={CheckProgress(text)}
-            controlHeight={26}
-            style={{ width: 81 }}
-            borderRadius={10}
-          >
-            {text.toLowerCase()}
-          </ButtonComponent>
-        </Space>
-      );
-    },
-  },
-  {
-    title: "Edit",
-    key: "edit",
-    dataIndex: "edit",
-    align: "center",
-    render: () => {
-      return (
-        <div className={"text-center"}>
-          <IconComponent
-            className={"text-xl text-indigo-500 border border-indigo-600"}
-            style={{
-              borderRadius: 5,
-              paddingLeft: 4,
-              paddingRight: 4,
-            }}
-            icon={<FormOutlined />}
-          />
-        </div>
-      );
-    },
-  },
-  {
-    title: "Delete",
-    key: "delete",
-    dataIndex: "delete",
-    align: "center",
-    render: () => {
-      return (
-        <div className={"text-center"}>
-          <IconComponent className={"w-7"} icon={<Icons type={"cross"} />} />
-        </div>
-      );
-    },
-  },
-];
-
-const data = [
-  {
-    category: "Student contract",
-    status: "active",
-  },
-  {
-    category: "Student contract",
-    status: "process",
-  },
-  {
-    category: "Student contract",
-    status: "close",
-  },
-];
-
 const File = () => {
   const { colorsObject } = useContext(ColorsContext);
+  const { columns, data } = FileCategoryModule();
   const [IsOpen, setIsOpen] = useState(false);
   const handleCategoryModal = () => setIsOpen((prev) => !prev);
 
@@ -165,11 +92,11 @@ const File = () => {
               <div className="space-y-2 5">
                 <div className="flex gap-3.5 items-center justify-between">
                   <ButtonComponent
-                    defaultBg={"#1890FF"}
-                    defaultHoverBg={"#1890FF"}
+                    defaultBg={colorsObject.info}
+                    defaultHoverBg={colorsObject.infoHover}
                     paddingInline={72}
-                    controlHeight={40}
-                    borderRadius={10}
+                    controlHeight={50}
+                    borderRadius={5}
                   >
                     Show more
                   </ButtonComponent>
@@ -182,11 +109,11 @@ const File = () => {
 
                 <div className="flex gap-3.5 items-center justify-between">
                   <ButtonComponent
-                    defaultBg={"#1890FF"}
-                    defaultHoverBg={"#1890FF"}
+                    defaultBg={colorsObject.info}
+                    defaultHoverBg={colorsObject.infoHover}
                     paddingInline={72}
-                    controlHeight={40}
-                    borderRadius={10}
+                    controlHeight={50}
+                    borderRadius={5}
                   >
                     Show more
                   </ButtonComponent>
@@ -200,11 +127,11 @@ const File = () => {
               <div className="space-y-2 5">
                 <div className="flex gap-3.5 items-center justify-between">
                   <ButtonComponent
-                    defaultBg={"#1890FF"}
-                    defaultHoverBg={"#1890FF"}
+                    defaultBg={colorsObject.info}
+                    defaultHoverBg={colorsObject.infoHover}
                     paddingInline={72}
-                    controlHeight={40}
-                    borderRadius={10}
+                    controlHeight={50}
+                    borderRadius={5}
                   >
                     Show more
                   </ButtonComponent>
@@ -231,20 +158,21 @@ const File = () => {
 
                 <div className="space-x-2 5">
                   <ButtonComponent
-                    defaultBg={"#24C18F"}
-                    defaultHoverBg={"#24C18F"}
+                    defaultBg={colorsObject.success}
+                    defaultHoverBg={colorsObject.successHover}
                     paddingInline={27}
-                    controlHeight={26}
+                    controlHeight={50}
                     borderRadius={5}
-                    onClick={handleCategoryModal}
+                    href={"/modals/management-file/new-category"}
+                    className={"inline-flex items-center"}
                   >
                     new category
                   </ButtonComponent>
                   <CustomSelect
                     colorBorder={colorsObject.info}
-                    style={{ width: 128, height: 26 }}
+                    style={{ width: 128 }}
                     selectorBg={colorsObject.info}
-                    className={`${FileStyle["File__select"]}`}
+                    className={`h-[50px] ${FileStyle["File__select"]}`}
                     value={"Filter by"}
                     options={[
                       {
@@ -263,7 +191,7 @@ const File = () => {
                   />
                 </div>
               </div>
-              <Table columns={columns} dataSource={data} pagination={false} />
+              <TableComponent columns={columns} data={data} />
             </div>
           </div>
           <div className={"space-y-2.5"}>
@@ -313,31 +241,15 @@ const File = () => {
                 </div>
               </div>
 
-              <div className="flex items-center justify-evenly">
-                <div className="w-36 ">
-                  <div
-                    className={`mb-2 relative ${FileStyle["Statistic__graph"]}`}
-                  >
-                    <CircularProgressbar
-                      value={0.33}
-                      maxValue={1}
-                      text={`${0.33 * 100}%`}
-                      strokeWidth={20}
-                      circleRatio={-1}
-                      styles={{
-                        text: {
-                          fill: colorsObject.black,
-                        },
-                        path: {
-                          stroke: "#24C18F",
-                        },
-                      }}
-                    />
-                  </div>
+              <div className="grid grid-cols-3 items-center">
+                <div className="">
+                  <FileChart
+                    colors={[colorsObject.secondary, colorsObject.success]}
+                  />
 
                   <Paragraph
                     className={
-                      "-mt-9 relative z-10 text-center text-base font-normal"
+                      "-mt-10 relative z-10 text-center text-base font-normal bg-white py-3"
                     }
                   >
                     Storage
@@ -346,38 +258,21 @@ const File = () => {
 
                 <ButtonComponent
                   defaultBg={colorsObject.info}
-                  defaultHoverBg={colorsObject.info}
-                  paddingInline={15}
-                  controlHeight={31}
-                  borderRadius={10}
+                  defaultHoverBg={colorsObject.infoHover}
+                  controlHeight={50}
+                  borderRadius={5}
                 >
                   Show more
                 </ButtonComponent>
 
-                <div className="w-36 ">
-                  <div
-                    className={`mb-2 relative ${FileStyle["Statistic__graph"]}`}
-                  >
-                    <CircularProgressbar
-                      value={0.33}
-                      maxValue={1}
-                      text={`${0.33 * 100}%`}
-                      strokeWidth={20}
-                      circleRatio={-1}
-                      styles={{
-                        text: {
-                          fill: colorsObject.black,
-                        },
-                        path: {
-                          stroke: colorsObject.danger,
-                        },
-                      }}
-                    />
-                  </div>
+                <div className="">
+                  <FileChart
+                    colors={[colorsObject.secondary, colorsObject.danger]}
+                  />
 
                   <Paragraph
                     className={
-                      "-mt-9 relative z-10 text-center text-base font-normal"
+                      "-mt-10 relative z-10 text-center text-base font-normal bg-white py-3"
                     }
                   >
                     Upload download
@@ -404,9 +299,9 @@ const File = () => {
 
                 <CustomSelect
                   colorBorder={colorsObject.info}
-                  style={{ width: 128, height: 26 }}
+                  style={{ width: 128 }}
                   selectorBg={colorsObject.info}
-                  className={`${FileStyle["File__select"]}`}
+                  className={`h-[50px] ${FileStyle["File__select"]}`}
                   value={"Filter by"}
                   options={[
                     {
@@ -489,7 +384,7 @@ const File = () => {
 
               <div className="text-end">
                 <ButtonComponent
-                  controlHeight={26}
+                  controlHeight={50}
                   defaultBg={colorsObject.main}
                   defaultHoverBg={colorsObject.main}
                   defaultColor={colorsObject.black}
@@ -505,217 +400,6 @@ const File = () => {
           </div>
         </div>
       </section>
-
-      {IsOpen && (
-        <Modal setIsOpen={setIsOpen}>
-          <div
-            className={classNames(
-              FileStyle["Modal__content-category"],
-              "bg-white py-7 px-12 w-full rounded-2xl overflow-scroll m-2.5",
-            )}
-          >
-            <Title
-              level={2}
-              fontSize={"text-indigo-600 text-4xl"}
-              fontWeightStrong={600}
-              titleMarginBottom={46}
-            >
-              Add new file category
-            </Title>
-
-            <form className="flex gap-5 flex-col">
-              <CustomInput
-                className={"border-indigo-700 border"}
-                classNames={
-                  "inline-flex justify-center h-10 items-center flex-row-reverse gap-5"
-                }
-                spanText={"Category name"}
-                spanClassName={"font-semibold flex-shrink-0"}
-                fontSize={"text-base"}
-                placeholder={"Category name"}
-              />
-
-              <label
-                className={"inline-flex justify-center items-center gap-5"}
-              >
-                <span
-                  className={
-                    "font-semibold text-end w-32 flex-shrink-0 text-base"
-                  }
-                >
-                  File status
-                </span>
-
-                <CustomSelect
-                  placeholder={"Select"}
-                  style={{ width: "100%" }}
-                  colorBorder={colorsObject.primary}
-                  className={"h-10"}
-                  options={[
-                    {
-                      value: "active",
-                      label: "active",
-                    },
-                  ]}
-                />
-              </label>
-
-              <label
-                className={"inline-flex justify-center items-center gap-5"}
-              >
-                <span
-                  className={
-                    "font-semibold text-end w-32 flex-shrink-0 text-base"
-                  }
-                >
-                  Packages:
-                </span>
-
-                <CustomTransfer
-                  dataSource={mockData}
-                  titles={["Source", "Target"]}
-                  colorBorder={colorsObject.primary}
-                  colorBgContainer={"transparent"}
-                  headerHeight={30}
-                  listHeight={200}
-                />
-              </label>
-
-              <label
-                className={"inline-flex justify-center items-center gap-5"}
-              >
-                <span
-                  className={
-                    "font-semibold text-end w-52 flex-shrink-0 text-base"
-                  }
-                >
-                  Signature link:
-                </span>
-
-                <textarea
-                  className={
-                    "border outline-0 border-indigo-700 p-5 rounded-2xl w-full min-h-[90px]"
-                  }
-                  placeholder={"text"}
-                ></textarea>
-              </label>
-
-              <label
-                className={"inline-flex justify-center items-center gap-5"}
-              >
-                <span
-                  className={
-                    "font-semibold text-end w-52 flex-shrink-0 text-base"
-                  }
-                >
-                  Note:
-                </span>
-
-                <textarea
-                  className={
-                    "border outline-0 border-indigo-700 p-5 rounded-2xl w-full min-h-[90px]"
-                  }
-                  placeholder={"text"}
-                ></textarea>
-              </label>
-
-              <div className={"grid grid-cols-2 gap-y-5 pt-8"}>
-                <label
-                  className={"inline-flex justify-center items-center gap-8"}
-                >
-                  <span
-                    className={
-                      "font-semibold text-end w-52 flex-shrink-0 text-base"
-                    }
-                  >
-                    Display on Student Portal:
-                  </span>
-                  <Switch
-                    style={{
-                      width: 50,
-                    }}
-                  />
-                </label>
-
-                <label
-                  className={"inline-flex justify-center items-center gap-8"}
-                >
-                  <span
-                    className={
-                      "font-semibold text-end w-52 flex-shrink-0 text-base"
-                    }
-                  >
-                    Must Be Uploaded to Student Account:
-                  </span>
-                  <Switch
-                    style={{
-                      width: 50,
-                    }}
-                  />
-                </label>
-
-                <label
-                  className={"inline-flex justify-center items-center gap-8"}
-                >
-                  <span
-                    className={
-                      "font-semibold text-end w-52 flex-shrink-0 text-base"
-                    }
-                  >
-                    Disallow files associated with category from displaying on
-                    Student Portal:
-                  </span>
-                  <Switch
-                    style={{
-                      width: 50,
-                    }}
-                  />
-                </label>
-
-                <label
-                  className={"inline-flex justify-center items-center gap-8"}
-                >
-                  <span
-                    className={
-                      "font-semibold text-end w-52 flex-shrink-0 text-base"
-                    }
-                  >
-                    Disallow files associated with this category from displaying
-                    on Instructor/Teacher Portal:
-                  </span>
-                  <Switch
-                    style={{
-                      width: 50,
-                    }}
-                  />
-                </label>
-              </div>
-
-              <div className={"text-center space-x-5"}>
-                <ButtonComponent
-                  defaultBg={colorsObject.secondary}
-                  defaultHoverBg={colorsObject.secondary}
-                  controlHeight={40}
-                  borderRadius={5}
-                  paddingInline={62}
-                  onClick={handleCategoryModal}
-                >
-                  Close
-                </ButtonComponent>
-                <ButtonComponent
-                  defaultBg={"#24C18F"}
-                  defaultHoverBg={"#24C18F"}
-                  controlHeight={40}
-                  borderRadius={5}
-                  paddingInline={62}
-                >
-                  Save
-                </ButtonComponent>
-              </div>
-            </form>
-          </div>
-        </Modal>
-      )}
     </Fragment>
   );
 };
