+++ frontend/src/pages/managment/management-spa.jsx
@@ -9,6 +9,7 @@ import { ModalContent } from "@/pages/managment/management-spa/modal/index.jsx";
 import ServiceStyle from "@/pages/managment/management.module.scss";
 import { StatusSelect } from "@/pages/managment/service/index.jsx";
 import { setActiveNav } from "@/modules/active-nav.jsx";
+import ManagementStyle from "@/pages/managment/management.module.scss";
 import { Pagination } from "antd";
 import { Fragment, useContext, useState } from "react";
 import { AiOutlineSearch } from "react-icons/ai";
@@ -50,7 +51,7 @@ const ManagementSpa = ({ page }) => {
 
         <div className="bg-white px-5 rounded-2xl">
           <div className="-mx-5">
-            <div className="space-x-7 px-5 border-b-2 border-b-gray-400">
+            <div className="space-x-7 px-5 border-b border-b-gray-400">
               <NavLink
                 to={"/management/single-page/location"}
                 className={setActiveNav}
@@ -79,80 +80,64 @@ const ManagementSpa = ({ page }) => {
           </div>
 
           <div className={"pt-5 pb-7"}>
-            <div>
-              <div className={"flex justify-between"}>
-                <form className={"flex gap-x-5 items-center"}>
-                  <label className={"relative"}>
-                    <CustomInput
-                      colorBorder={colorsObject.primary}
-                      placeholder={"Search"}
-                      className={`w-96 pl-12 pr-4 text-sm ${subpage === "quiz-report" && `inline-flex flex-row-reverse`} `}
-                    />
+            <div className={"flex justify-between"}>
+              <form className={"flex gap-x-5 items-center"}>
+                <label
+                  className={`relative h-[50px] rounded ${ManagementStyle["CheckModal__form-element__shadow"]}`}
+                >
+                  <CustomInput
+                    colorBorder={colorsObject.primary}
+                    placeholder={"Search"}
+                    classNames={"h-[50px]"}
+                    className={`w-96 pl-12 pr-4 text-sm ${subpage === "quiz-report" && `inline-flex flex-row-reverse`} `}
+                  />
 
-                    <span
-                      className={
-                        "absolute left-4 top-1/2 w-5 h-5 -translate-y-1/2 "
-                      }
-                    >
-                      <AiOutlineSearch />
-                    </span>
-                  </label>
-                  <ButtonComponent
-                    defaultBg={"#24C18F"}
-                    defaultHoverBg={"#24C18F"}
-                    paddingInline={26}
-                    controlHeight={40}
-                    borderRadius={5}
-                    className={"inline-flex items-center"}
-                    onClick={handleModal}
-                    href={
-                      subpage === "location"
-                        ? "/management/modal/location"
-                        : null
+                  <span
+                    className={
+                      "absolute left-4 top-1/2 w-5 h-5 -translate-y-1/2 "
                     }
                   >
-                    Add new
-                  </ButtonComponent>
+                    <AiOutlineSearch />
+                  </span>
+                </label>
+                <ButtonComponent
+                  defaultBg={colorsObject.success}
+                  defaultHoverBg={colorsObject.successHover}
+                  paddingInline={26}
+                  controlHeight={50}
+                  borderRadius={5}
+                  className={"inline-flex items-center"}
+                  href={`/modals/management/${subpage}`}
+                >
+                  Add new
+                </ButtonComponent>
 
-                  <CustomSelect
-                    value={"Status"}
-                    options={StatusSelect}
-                    style={{
-                      width: 122,
-                      height: 40,
-                    }}
-                    className={`${ServiceStyle["Service__select"]}`}
-                    colorBorder={colorsObject.info}
-                    selectorBg={colorsObject.info}
-                  />
-                </form>
-
-                <Pagination
-                  total={10}
-                  pageSize={1}
-                  current={CurrentPagination}
-                  onChange={handleChangePagination}
+                <CustomSelect
+                  value={"Status"}
+                  options={StatusSelect}
+                  style={{
+                    width: 122,
+                  }}
+                  className={`h-[50px] ${ServiceStyle["Service__select"]}`}
+                  colorBorder={colorsObject.info}
+                  selectorBg={colorsObject.info}
                 />
-              </div>
+              </form>
+
+              <Pagination
+                total={10}
+                pageSize={1}
+                current={CurrentPagination}
+                onChange={handleChangePagination}
+              />
             </div>
 
-            <div className="pt-5">
-              <div className={"-mx-5"}>
-                <ManagementSpaIndex />
-              </div>
+            <div className="pt-5 -mx-5">
+              <ManagementSpaIndex />
             </div>
           </div>
         </div>
       </section>
-      {IsOpen && (
-        <Modal setIsOpen={setIsOpen}>
-          <div
-            className={`bg-white rounded-2xl p-9 ${ModalStyle["Modal__content"]} overflow-y-scroll ${ServiceStyle["Modal__content"]}`}
-          >
-            <ModalContent page={subpage} />
-          </div>
-        </Modal>
-      )}
     </Fragment>
   );
 };
