+++ frontend/src/pages/managment/service/index.jsx
@@ -5,17 +5,19 @@ import Title from "@/components/title/index.jsx";
 import ColorsContext from "@/context/colors.jsx";
 import ModalStyle from "@/components/modal/modal.module.scss";
 import { Subpages } from "@/modules/subpages.jsx";
+import ManagementStyle from "@/pages/managment/management.module.scss";
 import {
   DiscountModalContent,
   FeesModalContent,
   MiscellaneousModalContent,
   ProductModalContent,
 } from "@/pages/managment/service/modal.jsx";
+import { setActiveNav } from "@/modules/active-nav.jsx";
 import { Pagination } from "antd";
 import { Fragment, useContext, useState } from "react";
 import { Helmet } from "react-helmet";
 import { AiOutlineSearch } from "react-icons/ai";
-import { NavLink, Outlet, useParams } from "react-router-dom";
+import { NavLink, useParams } from "react-router-dom";
 import ServiceStyle from "../management.module.scss";
 
 export const StatusSelect = [
@@ -37,41 +39,14 @@ export const StatusSelect = [
 ];
 
 const Service = ({ subpage }) => {
-  // use
   const { colorsObject } = useContext(ColorsContext);
-  const { title } = useParams();
-  const [CurrentPagination, setCurrentPagination] = useState(1);
-  const [IsOpen, setIsOpen] = useState(false);
 
-  // functions
+  const [CurrentPagination, setCurrentPagination] = useState(1);
+
   const handleChangePagination = (page) => {
     setCurrentPagination(page);
   };
 
-  const handleCloseModal = () => setIsOpen(false);
-
-  const handleAdd = () => setIsOpen(true);
-
-  const CheckPageForModal = (modal) => {
-    switch (modal) {
-      case "product":
-        return <ProductModalContent />;
-      case "fees":
-        return <FeesModalContent />;
-      case "discounts":
-        return <DiscountModalContent />;
-      case "miscellaneous":
-        return <MiscellaneousModalContent />;
-      default:
-        console.log(`not found content for ${modal} page`);
-    }
-  };
-
-  // Custom
-  const setActiveNav = ({ isActive }) =>
-    isActive
-      ? `${ServiceStyle["Tab__link-active"]} text-lg`
-      : "hover:text-indigo-500 text-lg text-gray-700";
   return (
     <Fragment>
       <Helmet>
@@ -96,10 +71,8 @@ const Service = ({ subpage }) => {
           Service
         </Title>
 
-        <div className={"p-5 bg-white rounded-3xl"}>
-          <div
-            className={"space-x-6 px-5 -mx-5 pb-5 border-b-2 border-b-gray-400"}
-          >
+        <div className={"px-5 bg-white rounded-3xl"}>
+          <div className={"space-x-6 px-5 -mx-5  border-b border-b-gray-400"}>
             <NavLink
               to={"/management/service/product"}
               className={setActiveNav}
@@ -146,14 +119,17 @@ const Service = ({ subpage }) => {
               Services (Packages)
             </NavLink>
           </div>
-          <div className={"pt-4 pb-7"} hidden={!title}>
+          <div className={"pt-4 pb-7"} hidden={!subpage}>
             <div className={"flex justify-between"}>
               <form className={"flex gap-x-5 items-center"}>
-                <label className={"relative"}>
+                <label
+                  className={`relative h-[50px] rounded ${ManagementStyle["CheckModal__form-element__shadow"]}`}
+                >
                   <CustomInput
                     colorBorder={colorsObject.primary}
                     placeholder={"Search"}
-                    className={`w-96 pl-12 pr-4 text-sm ${title === "quiz-report" && `inline-flex flex-row-reverse`} `}
+                    classNames={"h-[50px]"}
+                    className={`w-96 pl-12 pr-4 text-sm ${subpage === "quiz-report" && `inline-flex flex-row-reverse`} `}
                   />
 
                   <span
@@ -164,42 +140,38 @@ const Service = ({ subpage }) => {
                     <AiOutlineSearch />
                   </span>
                 </label>
-                {title !== "quiz-report" && (
+                {subpage !== "quiz-report" && (
                   <Fragment>
                     <ButtonComponent
-                      defaultBg={"#24C18F"}
-                      defaultHoverBg={"#24C18F"}
+                      defaultBg={colorsObject.success}
+                      defaultHoverBg={colorsObject.successHover}
+                      defaultBorderColor={colorsObject.success}
+                      defaultHoverBorderColor={colorsObject.successHover}
                       paddingInline={26}
-                      controlHeight={40}
+                      controlHeight={50}
                       borderRadius={5}
-                      onClick={title !== "packages" && handleAdd}
                       className={"inline-flex items-center"}
-                      href={
-                        title === "packages"
-                          ? "/management/modal/add-service"
-                          : null
-                      }
+                      href={`/modals/management-service/${subpage}`}
                     >
                       Add new
                     </ButtonComponent>
 
                     <CustomSelect
-                      value={"Status"}
+                      placeholder={"Status"}
                       options={StatusSelect}
                       style={{
                         width: 122,
-                        height: 40,
                       }}
-                      className={`${ServiceStyle["Service__select"]}`}
-                      colorBorder={"#1890FF"}
-                      selectorBg={"#1890FF"}
+                      className={`h-[50px] ${ServiceStyle["Service__select"]}`}
+                      colorBorder={colorsObject.info}
+                      selectorBg={colorsObject.info}
                     />
                   </Fragment>
                 )}
               </form>
 
               <div>
-                {title !== "quiz-report" && (
+                {subpage !== "quiz-report" && (
                   <Pagination
                     total={10}
                     pageSize={1}
@@ -209,47 +181,12 @@ const Service = ({ subpage }) => {
                 )}
               </div>
             </div>
-            <div className={"pt-5"}>
-              <div className={"-mx-5"}>
-                <Subpages page={subpage} />
-              </div>
+            <div className={"pt-5 -mx-5"}>
+              <Subpages page={subpage} />
             </div>
           </div>
         </div>
       </section>
-
-      {IsOpen && (
-        <Modal setIsOpen={setIsOpen}>
-          <div
-            className={`${ModalStyle["Modal__content"]} py-5 bg-white rounded-3xl ${title === "discounts" && "overflow-y-auto overflow-hidden"}`}
-          >
-            {CheckPageForModal(subpage)}
-
-            <div className={`text-center space-x-4`}>
-              <ButtonComponent
-                defaultBg={"#24C18F"}
-                defaultHoverBg={"#24C18F"}
-                paddingInline={64}
-                controlHeight={40}
-                onClick={handleCloseModal}
-                borderRadius={5}
-              >
-                Save
-              </ButtonComponent>
-              <ButtonComponent
-                controlHeight={39}
-                defaultBg={"#00000040"}
-                defaultHoverBg={"#00000040"}
-                paddingInline={60}
-                onClick={handleCloseModal}
-                borderRadius={5}
-              >
-                Close
-              </ButtonComponent>
-            </div>
-          </div>
-        </Modal>
-      )}
     </Fragment>
   );
 };
