+++ frontend/src/pages/managment/staff/index.jsx
@@ -1,8 +1,11 @@
 import ButtonComponent from "@/components/button/index.jsx";
 import { CustomInput, CustomSelect } from "@/components/form/index.jsx";
 import IconComponent from "@/components/icons/index.jsx";
+import TableComponent from "@/components/table/index.jsx";
 import Title, { Paragraph } from "@/components/title/index.jsx";
 import ColorsContext from "@/context/colors.jsx";
+import { StaffModule } from "@/modules/staff.jsx";
+import ManagementStyle from "@/pages/managment/management.module.scss";
 import ServiceStyle from "@/pages/managment/management.module.scss";
 import { StatusSelect } from "@/pages/managment/service/index.jsx";
 import { Pagination, Space, Table } from "antd";
@@ -15,182 +18,9 @@ import {
   FormOutlined,
 } from "@ant-design/icons";
 
-const CheckProgress = (status = "") => {
-  const { colorsObject } = useContext(ColorsContext);
-  switch (status.toLowerCase()) {
-    case "active":
-      return "#24C18F";
-    case "process":
-      return colorsObject.orange;
-    case "close":
-      return colorsObject.danger;
-    default:
-      return colorsObject.main;
-  }
-};
-
-const columns = [
-  {
-    title: "Last name",
-    dataIndex: "name",
-    key: "name",
-    render: (text) => {
-      return (
-        <Paragraph
-          className={"text-start"}
-          fontSize={"text-lg"}
-          fontWeightStrong={400}
-        >
-          {text}
-        </Paragraph>
-      );
-    },
-  },
-  {
-    title: "First name",
-    dataIndex: "firstName",
-    key: "firstName",
-    render: (text) => {
-      return (
-        <Paragraph
-          className={"text-start"}
-          fontSize={"text-lg"}
-          fontWeightStrong={400}
-        >
-          {text}
-        </Paragraph>
-      );
-    },
-  },
-  {
-    title: "Type",
-    dataIndex: "type",
-    key: "type",
-    render: (text) => {
-      return (
-        <Paragraph
-          className={"text-start"}
-          fontSize={"text-lg"}
-          fontWeightStrong={400}
-        >
-          {text}
-        </Paragraph>
-      );
-    },
-  },
-  {
-    title: "Cell phone",
-    dataIndex: "phone",
-    key: "phone",
-    render: (text) => {
-      return (
-        <Paragraph
-          className={"text-start"}
-          fontSize={"text-lg"}
-          fontWeightStrong={400}
-        >
-          {text}
-        </Paragraph>
-      );
-    },
-  },
-  {
-    title: "Status",
-    dataIndex: "status",
-    key: "status",
-    render: (text) => {
-      return (
-        <Space size={"middle"}>
-          <ButtonComponent
-            defaultBg={CheckProgress(text)}
-            defaultHoverBg={CheckProgress(text)}
-            controlHeight={30}
-            borderRadius={5}
-            style={{ width: "128px" }}
-          >
-            {text.toUpperCase()}
-          </ButtonComponent>
-        </Space>
-      );
-    },
-  },
-  {
-    title: "Action",
-    key: "action",
-    render: () => (
-      <Space size={"middle"}>
-        <IconComponent
-          className={"text-xl text-indigo-500 border border-indigo-600"}
-          style={{
-            borderRadius: 5,
-            paddingLeft: 4,
-            paddingRight: 4,
-          }}
-          icon={<FormOutlined />}
-        />
-
-        <IconComponent
-          className={"text-xl text-red-600 border border-indigo-600"}
-          style={{
-            borderRadius: 5,
-            paddingLeft: 4,
-            paddingRight: 4,
-          }}
-          icon={<DeleteOutlined />}
-        />
-
-        <IconComponent
-          className={"text-xl text-indigo-500 border border-indigo-600"}
-          style={{
-            borderRadius: 5,
-            paddingLeft: 4,
-            paddingRight: 4,
-          }}
-          icon={<ExportOutlined />}
-        />
-      </Space>
-    ),
-  },
-];
-
-const data = [
-  {
-    key: "1",
-    name: "Aminov",
-    firstName: "Makhsud",
-    type: "Owner",
-    subtype: "Teen BTW",
-    hours: 1,
-    observation: 1,
-    status: "Active",
-    phone: "(513)837-5128",
-  },
-  {
-    key: "2",
-    name: "Aminov",
-    firstName: "Makhsud",
-    type: "Owner",
-    subtype: "Teen BTW",
-    hours: 1,
-    observation: 1,
-    status: "Close",
-    phone: "(513)837-5128",
-  },
-  {
-    key: "3",
-    name: "Aminov",
-    firstName: "Makhsud",
-    type: "Owner",
-    subtype: "Teen BTW",
-    hours: 1,
-    observation: 1,
-    status: "process",
-    phone: "(513)837-5128",
-  },
-];
-
 const Staff = () => {
   const { colorsObject } = useContext(ColorsContext);
+  const { data, columns } = StaffModule();
   const [CurrentPagination, setCurrentPagination] = useState(1);
   const handleChangePagination = (page) => {
     setCurrentPagination(page);
@@ -223,30 +53,31 @@ const Staff = () => {
         <div className={"py-6 px-5 bg-white rounded-2xl shadow-lg"}>
           <div className={"flex justify-between items-center mb-5"}>
             <form className={"flex gap-x-5 items-center"}>
-              <label className={"relative shadow-lg"}>
+              <label
+                className={`relative h-[50px] rounded ${ManagementStyle["CheckModal__form-element__shadow"]}`}
+              >
                 <CustomInput
                   colorBorder={colorsObject.primary}
                   placeholder={"Search"}
-                  className={`w-96 pl-12 pr-4 text-sm inline-flex flex-row-reverse `}
+                  classNames={"h-[50px]"}
+                  className={`w-96 pl-12 pr-4 text-sm`}
                 />
 
                 <span
-                  className={
-                    "absolute left-4 top-1/2 w-5 h-5 -translate-y-1/2 "
-                  }
+                  className={"absolute left-4 top-1/2 w-5 h-5 -translate-y-1/2"}
                 >
                   <AiOutlineSearch />
                 </span>
               </label>
               <Fragment>
                 <ButtonComponent
-                  defaultBg={"#24C18F"}
-                  defaultHoverBg={"#24C18F"}
+                  defaultBg={colorsObject.success}
+                  defaultHoverBg={colorsObject.successHover}
                   paddingInline={26}
-                  controlHeight={40}
+                  controlHeight={50}
                   borderRadius={5}
                   className={"inline-flex items-center"}
-                  href={"/management/modal/add-staff"}
+                  href={"/modals/staff/add-staff"}
                 >
                   Add new
                 </ButtonComponent>
@@ -256,9 +87,8 @@ const Staff = () => {
                   options={StatusSelect}
                   style={{
                     width: 122,
-                    height: 40,
                   }}
-                  className={`${ServiceStyle["Service__select"]}`}
+                  className={`h-[50px] ${ServiceStyle["Service__select"]}`}
                   colorBorder={"#1890FF"}
                   selectorBg={"#1890FF"}
                 />
@@ -273,7 +103,7 @@ const Staff = () => {
             />
           </div>
 
-          <Table columns={columns} dataSource={data} pagination={false} />
+          <TableComponent columns={columns} data={data} />
         </div>
       </Fragment>
     </Fragment>
